Refactor(App): Extract functions and components, fix lint errors
This commit refactors the main `App.jsx` component to improve readability and reduce its size. The changes follow a progressive approach to decentralize logic and UI.

Key changes include:

- **Function Extraction**: Several helper functions were moved from `App.jsx` to a new utility file at `src/lib/helpers.js`. This includes logic for CSV handling, IA responses, color palette generation, and HTML export. `App.jsx` was updated to import and use these functions.

- **Component Extraction**: The JSX and state management for the "Campanha" (Campaign) step, which was a large and complex part of `App.jsx`, has been extracted into a new, dedicated component: `src/components/Campaign.jsx`. `App.jsx` now renders this component and passes the required data and handlers as props.

- **Linting and Cleanup**: Fixed all outstanding linter errors and warnings within `App.jsx`. This involved:
    - Removing dead code and unused state variables related to a password-protection feature that had been previously moved to the `SetupModal` component.
    - Correcting a `useEffect` dependency array to prevent potential bugs.
    - Replacing calls to undefined functions with the correct handlers, improving the stability of the component.

These changes make `App.jsx` significantly smaller and easier to maintain, primarily acting as a top-level state manager and router for the application's different steps.
