+++ src/App.jsx
@@ -62,7 +62,7 @@ import {
   FormatBold,
   Visibility,
   Grid3x3,
-  Campaign,
+  Campaign as CampaignIcon,
   AspectRatio,
   Language,
   Publish,
@@ -77,8 +77,6 @@ import { ThemeProvider, createTheme } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { Toaster, toast } from 'sonner';
 
-import { saveCredentialsToFile, loadCredentialsFromFile } from './utils/credentialsManager';
-import PasswordDialog from './components/PasswordDialog';
 import FieldPositioner from './components/FieldPositioner';
 import FormattingPanel from './components/FormattingPanel';
 import FormattingDrawer from './components/FormattingDrawer';
@@ -100,6 +98,7 @@ import { stripHtml } from './lib/utils';
 import './App.css';
 import LoadingDialog from './components/LoadingDialog';
 import TextEditorDialog from './components/TextEditorDialog';
+import Campaign from './components/Campaign';
 
 // Temas atualizados com gradientes e cores modernas
 const lightTheme = createTheme({
@@ -259,10 +258,6 @@ function App() {
   const [includeEmpresa, setIncludeEmpresa] = useState(true);
   const [showSetupModal, setShowSetupModal] = useState(false);
   const [showCampaignStandardsModal, setShowCampaignStandardsModal] = useState(false);
-  const [showPasswordDialog, setShowPasswordDialog] = useState(false);
-  const [passwordDialogAction, setPasswordDialogAction] = useState(null); // 'save' or 'load'
-  const [credentialsPassword, setCredentialsPassword] = useState('');
-
   const handleGenerateColorPalette = async (briefing) => {
     const apiKey = getGeminiApiKey();
     if (!apiKey) {
@@ -320,7 +315,7 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
     }
   };
 
-  const saveStateToSessionStorage = async () => {
+  const saveStateToSessionStorage = useCallback(async () => {
     const blobToBase64 = (blob) => {
       return new Promise((resolve, reject) => {
         const reader = new FileReader();
@@ -411,7 +406,30 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
       promptText,
     };
     sessionStorage.setItem('appState', JSON.stringify(stateToSave));
-  };
+  }, [
+    activeStep,
+    csvData,
+    csvHeaders,
+    backgroundImage,
+    colorPalette,
+    fieldPositions,
+    fieldStyles,
+    displayedImageSize,
+    originalImageSize,
+    generatedVideosData,
+    problema,
+    solucao,
+    campaignContent,
+    persona,
+    autor,
+    instrucoes,
+    formato,
+    aspectRatio,
+    generatedImageUrl,
+    conteudoMedio,
+    conteudoPequeno,
+    promptText,
+  ]);
 
   useEffect(() => {
     const loadStateFromSession = async () => {
@@ -489,7 +507,7 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
     if (generatedVideosData.length > 0) {
       saveStateToSessionStorage();
     }
-  }, [generatedVideosData]);
+  }, [generatedVideosData, saveStateToSessionStorage]);
 
 
   // Estados para a Campanha
@@ -536,7 +554,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
   const fileInputRef = useRef(null);
   const imageInputRef = useRef(null);
   const loadStateInputRef = useRef(null);
-  const loadCredentialsInputRef = useRef(null);
 
   useEffect(() => {
     localStorage.setItem('darkMode', JSON.stringify(darkMode));
@@ -597,7 +614,7 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
             });
 
             // Abre o modal para mostrar que a conexão foi bem-sucedida
-            setShowLinkedinAuthModal(true);
+            toast.success('LinkedIn conectado com sucesso!');
           } else {
             console.error('Erro ao obter o token de acesso do LinkedIn:', data);
             // Tratar erro, talvez mostrar mensagem
@@ -1321,40 +1338,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
   const handleSaveTemplateClick = () => {
     handleMenuClose();
     handleSaveState();
-  };
-
-  const handlePasswordConfirm = async (password) => {
-    setShowPasswordDialog(false);
-    if (passwordDialogAction === 'save') {
-      try {
-        await saveCredentialsToFile(password);
-        toast.success('Arquivo de credenciais salvo com sucesso!');
-      } catch (error) {
-        toast.error(`Erro ao salvar credenciais: ${error.message}`);
-      }
-    } else if (passwordDialogAction === 'load') {
-      setCredentialsPassword(password);
-      loadCredentialsInputRef.current.click();
-    }
-  };
-
-  const handleLoadCredentialsFileChange = async (event) => {
-    const file = event.target.files[0];
-    if (file && credentialsPassword) {
-      try {
-        await loadCredentialsFromFile(file, credentialsPassword);
-        toast.success('Credenciais carregadas com sucesso! A página será recarregada.');
-        setTimeout(() => window.location.reload(), 2000);
-      } catch (error) {
-        toast.error(`Erro ao carregar credenciais: ${error.message}`);
-      } finally {
-        setCredentialsPassword('');
-        // Reset file input
-        if (loadCredentialsInputRef.current) {
-          loadCredentialsInputRef.current.value = '';
-        }
-      }
-    }
   };
 
   const handleExportCSV = () => {
@@ -2232,13 +2215,6 @@ Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` co
                 onChange={handleLoadStateFromFile}
                 ref={loadStateInputRef}
               />
-              <input
-                type="file"
-                hidden
-                accept=".midiatorsetup"
-                onChange={handleLoadCredentialsFileChange}
-                ref={loadCredentialsInputRef}
-              />
             </Box>
           </Toolbar>
         </AppBar>
@@ -2351,282 +2327,41 @@ Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` co
         >
           {/* Passo 0: Campanha */}
           {activeStep === 0 && (
-            <Card>
-              <CardContent sx={{ p: { xs: 1.5, sm: 2, md: 4 } }}>
-                <Typography variant="h5" gutterBottom sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 3 }}>
-                  <Campaign />
-                  {steps[0].label}
-                </Typography>
-                <Grid container spacing={3}>
-                  <Grid item xs={12} md={6}>
-                    <TextField
-                      label="Problema"
-                      multiline
-                      rows={4}
-                      value={problema}
-                      onChange={(e) => setProblema(e.target.value)}
-                      variant="outlined"
-                      fullWidth
-                      placeholder="Descreva o problema que sua campanha busca resolver."
-                      disabled={campaignContent !== null}
-                    />
-                  </Grid>
-                  <Grid item xs={12} md={6}>
-                    <TextField
-                      label="Solução"
-                      multiline
-                      rows={4}
-                      value={solucao}
-                      onChange={(e) => setSolucao(e.target.value)}
-                      variant="outlined"
-                      fullWidth
-                      placeholder="Descreva a solução que sua campanha oferece."
-                      disabled={campaignContent !== null}
-                    />
-                  </Grid>
-                  <Grid item xs={12} md={6}>
-                    <TextField
-                        label="Quantidade de Posts de Follow-up"
-                        type="number"
-                        value={followupPostsQuantity}
-                        onChange={(e) => setFollowupPostsQuantity(parseInt(e.target.value, 10))}
-                        fullWidth
-                        variant="outlined"
-                        disabled={campaignContent !== null}
-                        InputProps={{ inputProps: { min: 1, max: 10 } }}
-                    />
-                  </Grid>
-                  <Grid item xs={12} md={6}>
-                    <FormControl fullWidth variant="outlined" disabled={campaignContent !== null}>
-                      <InputLabel id="aspect-ratio-label">Razão de Aspecto</InputLabel>
-                      <Select
-                        labelId="aspect-ratio-label"
-                        value={aspectRatio}
-                        onChange={(e) => setAspectRatio(e.target.value)}
-                        label="Razão de Aspecto"
-                      >
-                        <MenuItem value="1:1">Quadrado (1:1)</MenuItem>
-                        <MenuItem value="4:5">Retrato (4:5)</MenuItem>
-                        <MenuItem value="16:9">Paisagem (16:9)</MenuItem>
-                      </Select>
-                    </FormControl>
-                  </Grid>
-                </Grid>
-                <Box sx={{ display: 'flex', justifyContent: 'center', mt: 3, gap: 2 }}>
-                  <Button
-                    variant="contained"
-                    size="large"
-                    onClick={() => handleGenerateCampaignContent(false)}
-                    disabled={!problema.trim() || !solucao.trim() || isGeneratingCampaign || campaignContent !== null}
-                  >
-                    {isGeneratingCampaign ? 'Gerando...' : 'Gerar Tudo com IA'}
-                  </Button>
-                  {campaignContent && (
-                    <Button
-                      variant="outlined"
-                      size="large"
-                      onClick={handleResetCampaign}
-                    >
-                      Resetar
-                    </Button>
-                  )}
-                </Box>
-
-                {campaignContent && (
-                  <Box sx={{ mt: 4 }}>
-                    <Box sx={{ display: 'flex', justifyContent: 'flex-end', mb: 2 }}>
-                      <Button
-                        variant="outlined"
-                        onClick={handleExportHtml}
-                      >
-                        Exportar como HTML
-                      </Button>
-                    </Box>
-                    <Typography variant="h6" gutterBottom>Conteúdo Gerado</Typography>
-                    <Grid container spacing={2}>
-                      <Grid item xs={12}>
-                        <TextField
-                          label="Título"
-                          value={campaignContent.titulo}
-                          onChange={(e) => setCampaignContent({ ...campaignContent, titulo: e.target.value })}
-                          variant="outlined"
-                          fullWidth
-                        />
-                      </Grid>
-                      <Grid item xs={12} sx={{ display: 'flex', alignItems: 'center', gap: 2 }}>
-                        <TextField
-                          label="Conteúdo"
-                          multiline
-                          rows={4}
-                          value={campaignContent.conteudo}
-                          onClick={() => setEditingField('conteudo')}
-                          readOnly
-                          variant="outlined"
-                          fullWidth
-                          sx={{ cursor: 'pointer' }}
-                        />
-                        <Button onClick={() => handleGenerateCampaignContent(true)} disabled={isGeneratingCampaign}>Gerar</Button>
-                      </Grid>
-
-                      <Grid item xs={12} sx={{ display: 'flex', alignItems: 'center', gap: 2 }}>
-                        <TextField
-                          label="Conteúdo Médio (máx. 1800 caracteres)"
-                          multiline
-                          rows={2}
-                          value={conteudoMedio}
-                          onChange={(e) => setConteudoMedio(e.target.value)}
-                          variant="outlined"
-                          fullWidth
-                        />
-                        <Button onClick={() => handleGenerateSummary(1800)} disabled={isGeneratingSummaryMedio || !campaignContent}>
-                          {isGeneratingSummaryMedio ? 'Gerando...' : 'Gerar'}
-                        </Button>
-                      </Grid>
-
-                      <Grid item xs={12} sx={{ display: 'flex', alignItems: 'center', gap: 2 }}>
-                        <TextField
-                          label="Conteúdo Pequeno (máx. 130 caracteres)"
-                          multiline
-                          rows={1}
-                          value={conteudoPequeno}
-                          onChange={(e) => setConteudoPequeno(e.target.value)}
-                          variant="outlined"
-                          fullWidth
-                        />
-                         <Button onClick={() => handleGenerateSummary(130)} disabled={isGeneratingSummaryPequeno || !campaignContent}>
-                          {isGeneratingSummaryPequeno ? 'Gerando...' : 'Gerar'}
-                        </Button>
-                      </Grid>
-
-                      <Grid item xs={12} sx={{ display: 'flex', alignItems: 'center', gap: 2 }}>
-                        <TextField
-                          label="Conteúdo Formatado (HTML)"
-                          multiline
-                          rows={3}
-                          value={conteudoFormatado}
-                          onClick={() => setEditingField('conteudoFormatado')}
-                          readOnly
-                          variant="outlined"
-                          fullWidth
-                          sx={{ cursor: 'pointer' }}
-                        />
-                         <Button onClick={() => handleGenerateFormattedContent()} disabled={isGeneratingConteudoFormatado || !campaignContent}>
-                          {isGeneratingConteudoFormatado ? 'Gerando...' : 'Gerar'}
-                        </Button>
-                      </Grid>
-
-                      <Grid item xs={12}>
-                        <TextField
-                          label="CTA (Chamada para Ação)"
-                          multiline
-                          rows={2}
-                          value={campaignContent.cta}
-                          onClick={() => setEditingField('cta')}
-                          readOnly
-                          variant="outlined"
-                          fullWidth
-                          sx={{ cursor: 'pointer' }}
-                        />
-                      </Grid>
-                      <Grid item xs={12}>
-                        <Typography variant="subtitle2" gutterBottom>Hashtags</Typography>
-                        <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 1 }}>
-                          {campaignContent.hashtags.map((tag, index) => (
-                            <Chip
-                              key={index}
-                              label={tag}
-                              onDelete={() => {
-                                const newHashtags = [...campaignContent.hashtags];
-                                newHashtags.splice(index, 1);
-                                setCampaignContent({ ...campaignContent, hashtags: newHashtags });
-                              }}
-                            />
-                          ))}
-                        </Box>
-                        <Box sx={{ display: 'flex', gap: 1, mt: 1 }}>
-                          <TextField
-                            label="Nova Hashtag"
-                            size="small"
-                            onKeyDown={(e) => {
-                              if (e.key === 'Enter' && e.target.value.trim() !== '') {
-                                e.preventDefault();
-                                setCampaignContent({ ...campaignContent, hashtags: [...campaignContent.hashtags, e.target.value.trim()] });
-                                e.target.value = '';
-                              }
-                            }}
-                          />
-                           <Button onClick={() => {
-                              const newTag = document.querySelector('input[label="Nova Hashtag"]').value.trim();
-                              if (newTag) {
-                                setCampaignContent({ ...campaignContent, hashtags: [...campaignContent.hashtags, newTag] });
-                                document.querySelector('input[label="Nova Hashtag"]').value = '';
-                              }
-                           }}>Adicionar</Button>
-                        </Box>
-                      </Grid>
-                    </Grid>
-                  </Box>
-                )}
-
-                {isGeneratingFollowup && (
-                  <Box sx={{ mt: 4, textAlign: 'center' }}>
-                    <Typography variant="h6" gutterBottom>Gerando Posts de Follow-up...</Typography>
-                  </Box>
-                )}
-
-                {followupPosts.length > 0 && !isGeneratingFollowup && (
-                  <Box sx={{ mt: 4 }}>
-                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
-                      <Typography variant="h6" gutterBottom>Posts de Follow-up Gerados</Typography>
-                      <Button onClick={() => handleGenerateFollowupPosts()} disabled={isGeneratingFollowup}>
-                        {isGeneratingFollowup ? 'Gerando...' : 'Regenerar Posts'}
-                      </Button>
-                    </Box>
-                    {followupPosts.map((post, index) => (
-                      <Accordion key={index}>
-                        <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                          <Typography>Post {post.post_numero}: {post.tipo_gancho}</Typography>
-                        </AccordionSummary>
-                        <AccordionDetails>
-                          <Typography variant="body2" sx={{ whiteSpace: 'pre-wrap' }}>
-                            {post.conteudo}
-                          </Typography>
-                          <Typography variant="caption" color="textSecondary" sx={{ mt: 1, display: 'block' }}>
-                            CTA: {post.cta}
-                          </Typography>
-                          <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 0.5, mt: 1 }}>
-                            {post.hashtags_sugeridas.map((tag, i) => (
-                              <Chip key={i} label={tag} size="small" />
-                            ))}
-                          </Box>
-                        </AccordionDetails>
-                      </Accordion>
-                    ))}
-                  </Box>
-                )}
-
-                {isGeneratingImage && (
-                  <Box sx={{ mt: 4, textAlign: 'center' }}>
-                    <Typography variant="h6" gutterBottom>
-                      Gerando Imagem...
-                    </Typography>
-                    {/* Pode adicionar um componente de loading mais elaborado aqui */}
-                  </Box>
-                )}
-
-                {generatedImageUrl && !isGeneratingImage && (
-                  <Box sx={{ mt: 4 }}>
-                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                      <Typography variant="h6" gutterBottom>Imagem Gerada</Typography>
-                      <Button onClick={handleGenerateImage} disabled={isGeneratingImage}>
-                        {isGeneratingImage ? 'Gerando...' : 'Regenerar Imagem'}
-                      </Button>
-                    </Box>
-                    <img src={generatedImageUrl} alt="Imagem gerada pela IA" style={{ maxWidth: '100%', borderRadius: '8px', mt: 2 }} />
-                  </Box>
-                )}
-              </CardContent>
-            </Card>
+            <Campaign
+              steps={steps}
+              problema={problema}
+              setProblema={setProblema}
+              solucao={solucao}
+              setSolucao={setSolucao}
+              followupPostsQuantity={followupPostsQuantity}
+              setFollowupPostsQuantity={setFollowupPostsQuantity}
+              aspectRatio={aspectRatio}
+              setAspectRatio={setAspectRatio}
+              isGeneratingCampaign={isGeneratingCampaign}
+              campaignContent={campaignContent}
+              handleGenerateCampaignContent={handleGenerateCampaignContent}
+              handleResetCampaign={handleResetCampaign}
+              handleExportHtml={handleExportHtml}
+              editingField={editingField}
+              setEditingField={setEditingField}
+              conteudoMedio={conteudoMedio}
+              setConteudoMedio={setConteudoMedio}
+              isGeneratingSummaryMedio={isGeneratingSummaryMedio}
+              handleGenerateSummary={handleGenerateSummary}
+              conteudoPequeno={conteudoPequeno}
+              setConteudoPequeno={setConteudoPequeno}
+              isGeneratingSummaryPequeno={isGeneratingSummaryPequeno}
+              conteudoFormatado={conteudoFormatado}
+              isGeneratingConteudoFormatado={isGeneratingConteudoFormatado}
+              handleGenerateFormattedContent={handleGenerateFormattedContent}
+              followupPosts={followupPosts}
+              isGeneratingFollowup={isGeneratingFollowup}
+              handleGenerateFollowupPosts={handleGenerateFollowupPosts}
+              generatedImageUrl={generatedImageUrl}
+              isGeneratingImage={isGeneratingImage}
+              handleGenerateImage={handleGenerateImage}
+              setCampaignContent={setCampaignContent}
+            />
           )}
 
           {/* Passo 1: Definir Dados Iniciais */}
@@ -2766,7 +2501,7 @@ Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` co
                     {!getGeminiApiKey() && (
                       <Alert severity="warning" sx={{ mb: 2, width: '100%', maxWidth: '500px' }}>
                         Chave da API Gemini não configurada.
-                        <MuiLink component="button" variant="body2" onClick={() => setShowGeminiAuthModal(true)} sx={{ ml: 1 }}>
+                        <MuiLink component="button" variant="body2" onClick={() => setShowSetupModal(true)} sx={{ ml: 1 }}>
                           Configurar Chave Gemini
                         </MuiLink>
                       </Alert>
@@ -3139,17 +2874,6 @@ Lembre-se: Sua resposta final deve conter APENAS o bloco \`\`\`csv ... \`\`\` co
           }
         }}
         onClose={() => setEditingField(null)}
-      />
-      <PasswordDialog
-        open={showPasswordDialog}
-        onClose={() => setShowPasswordDialog(false)}
-        onConfirm={handlePasswordConfirm}
-        title={passwordDialogAction === 'save' ? 'Salvar Credenciais' : 'Carregar Credenciais'}
-        description={
-          passwordDialogAction === 'save'
-            ? 'Digite uma senha para criptografar o arquivo de credenciais.'
-            : 'Digite a senha para descriptografar o arquivo de credenciais.'
-        }
       />
       {isMobile && activeStep === 4 && (
         <>
